starbook/Auth/api/viewsets.py
from django.contrib.auth.models import User
from Auth.api.serializers import UserSerializer
from rest_framework import viewsets

class UserViewSet(viewsets.ModelViewSet):
    serializer_class = UserSerializer
    http_method_names = ['post']